feat(groth16): group operations
